+++ spec/api/v1/characters_controller_spec.rb
@@ -1,4 +1,31 @@
 RSpec.describe 'Characters API' do
+  describe 'GET#index' do
+    it_behaves_like 'API auth without token'
+    it_behaves_like 'API auth with invalid token'
+    it_behaves_like 'API auth unconfirmed'
+
+    context 'with valid user token in params' do
+      let!(:user) { create :user }
+      let(:access_token) { JwtService.new.json_response(user: user)[:access_token] }
+      let!(:character) { create :character, user: user }
+      before { get '/api/v1/characters.json', params: { access_token: access_token } }
+
+      it 'returns status 200' do
+        expect(response.status).to eq 200
+      end
+
+      %w[id name fraction_name world_id fraction_id].each do |attr|
+        it "and contains character #{attr}" do
+          expect(response.body).to have_json_path("characters/0/#{attr}")
+        end
+      end
+    end
+
+    def do_request(headers = {})
+      get '/api/v1/characters.json', headers: headers
+    end
+  end
+
   describe 'GET#show' do
     it_behaves_like 'API auth without token'
     it_behaves_like 'API auth with invalid token'